backend/jrats_project/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
redis==6.4.0
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
